% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareProteinInteractions.R
\name{compareProteinInteractions}
\alias{compareProteinInteractions}
\title{Compare the Interactions of 2 Proteins, Using a Venn Diagram}
\usage{
compareProteinInteractions(proteinOneID, proteinTwoID = NULL)
}
\arguments{
\item{proteinOneID}{the ID of a protein}

\item{proteinTwoID}{the optional ID of a second protein, will default to SARS-CoV-2 spike protein}
}
\description{
Compare the Interactions of 2 Proteins, Using a Venn Diagram
}
\references{
Gao, C.-H. (2021, October 22). A 'ggplot2' implement of Venn diagram. The Comprehensive R Archive Network. Retrieved from https://cran.r-project.org/web/packages/ggVennDiagram/.
}
