% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProteinInteractionGraph.R
\name{getProteinInteractionData}
\alias{getProteinInteractionData}
\title{Returns a dataframe of all possible protein-protein interactions
using a call to the IMEX database. No graphical output.}
\usage{
getProteinInteractionData(proteinId)
}
\arguments{
\item{proteinId}{the protein ID}
}
\description{
Returns a dataframe of all possible protein-protein interactions
using a call to the IMEX database. No graphical output.
}
\references{
Altfeld, J. (2021, October 24). Error handling in R with trycatchlog: Catching, logging, post-mortem analysis. CRAN. Retrieved December 9, 2021, from https://cran.r-project.org/web/packages/tryCatchLog/vignettes/tryCatchLog-intro.html.
Protein interaction data tools: Retrieving data from intact and most other databases using PSICQUIC service. PItools. (n.d.). Retrieved from https://vitkl.github.io/PItools/index.html.
}
