% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareProteinInteractions.R
\name{getUniqueProteins}
\alias{getUniqueProteins}
\title{Helper function for compareProteinInteractions().
Condenses a large protein dataframe into unique protein interactions}
\usage{
getUniqueProteins(proteinData)
}
\arguments{
\item{proteinData}{a dataframe of protein data from getProteinInteractionData()}
}
\value{
a list of unique proteins that this protein interacts with
}
\description{
Helper function for compareProteinInteractions().
Condenses a large protein dataframe into unique protein interactions
}
