% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseSpike.R
\name{pairwiseSpike}
\alias{pairwiseSpike}
\title{Pairwise compare 2 spike protines}
\usage{
pairwiseSpike(spike1, spike2, scoreCheck)
}
\arguments{
\item{spike1}{A spike sequence from a FAFSA file.}

\item{spike2}{A spike sequence from a FAFSA file.}

\item{scoreCheck}{Bool}
}
\value{
the score and comparison of the 2 proteins.
}
\description{
Pairwise compare 2 spike protines
}
