% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVSD.R
\name{runVSD}
\alias{runVSD}
\title{VSD Launch script
Launches a shiny app for VSD}
\usage{
runVSD()
}
\value{
opens a shiny page
}
\description{
VSD Launch script
Launches a shiny app for VSD
}
\references{
Dean Attali. (2015, December 7). Building shiny apps - an interactive tutorial. Dean Attali. Retrieved December 9, 2021, from https://deanattali.com/blog/building-shiny-apps-tutorial/#:~:text=You\%20can\%20also\%20create\%20a,R\%20\%2B\%20server.
}
