% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareProteinInteractions.R
\name{compareProteinInteractions}
\alias{compareProteinInteractions}
\title{Compare the Interactions of 2 Proteins, Using a Venn Diagram}
\usage{
compareProteinInteractions(proteinOneID, proteinTwoID = NULL)
}
\arguments{
\item{proteinOneID}{the ID of a protein}

\item{proteinTwoID}{the optional ID of a second protein, will default to SARS-CoV-2 spike protein}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Compare the Interactions of 2 Proteins, Using a Venn Diagram
}
\examples{
compareProteinInteractions("P0DTC2","G1PXH7")
compareProteinInteractions("G1PXH7")
}
