% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProteinInteractionGraph.R
\name{createProteinInteractionGraph}
\alias{createProteinInteractionGraph}
\title{Creates and outputs a Protein-Protein Interaction Graph}
\usage{
createProteinInteractionGraph(proteinId)
}
\arguments{
\item{proteinId}{the protein ID}
}
\description{
Creates and outputs a Protein-Protein Interaction Graph
}
\examples{
createProteinInteractionGraph("G1PXH7")
}
