% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProteinInteractionGraph.R
\name{getProteinInteractionData}
\alias{getProteinInteractionData}
\title{Returns a dataframe of all possible protein-protein interactions
using a call to the IMEX database. No graphical output.}
\usage{
getProteinInteractionData(proteinId)
}
\arguments{
\item{proteinId}{the protein ID}
}
\description{
Returns a dataframe of all possible protein-protein interactions
using a call to the IMEX database. No graphical output.
}
\examples{
getProteinInteractionData("G1PXH7")
}
